dryed it up, moved selector into tile
